src/requirements.txt
requests
pyyaml
pushbullet.py
webbrowser